+++ components/resolved_ts/tests/integrations/test_sst_lease.rs
@@ -27,7 +27,8 @@ fn test_lease_block_resolved_ts() {
     meta.set_region_id(region.id);
     meta.set_region_epoch(region.get_region_epoch().clone());
     // The resolved-ts won't be updated from acquiring sst lease till ingest sst.
-    suite.must_acquire_sst_lease(region.id, &meta, Duration::MAX);
+    let import = suite.get_import_client(region.id);
+    must_acquire_sst_lease(import, &meta, Duration::MAX);
     // The resolved-ts must be less than the latest ts.
     let sst_commit_ts = block_on(suite.cluster.pd_client.get_tso()).unwrap();
     let blocked_rts = suite.region_resolved_ts(region.id).unwrap();
@@ -37,12 +38,15 @@ fn test_lease_block_resolved_ts() {
         (sst_commit_ts, blocked_rts)
     );
 
-    let resp = suite.upload_sst(region.id, &meta, &data).unwrap();
+    let import = suite.get_import_client(region.id);
+    let resp = send_upload_sst(import, &meta, &data).unwrap();
     assert!(!resp.has_error(), "{:?}", resp);
     // The resolved-ts must be blocked.
     suite.must_get_rts(region.id, blocked_rts);
 
-    suite.must_ingest_sst(region.id, meta);
+    let ctx = suite.get_context(region.id);
+    let import = suite.get_import_client(region.id);
+    must_ingest_sst(import, ctx, meta);
 
     // The resolved-ts must advance after ingest sst.
     suite.must_get_rts_ge(region.id, blocked_rts);
@@ -67,17 +71,20 @@ fn test_lease_release_unblock_resolved_ts() {
     meta.set_region_id(region.id);
     meta.set_region_epoch(region.get_region_epoch().clone());
     // The resolved-ts won't be updated from acquiring sst lease till ingest sst.
-    suite.must_acquire_sst_lease(region.id, &meta, Duration::MAX);
+    let import = suite.get_import_client(region.id);
+    must_acquire_sst_lease(import, &meta, Duration::MAX);
     let blocked_rts1 = suite.region_resolved_ts(region.id).unwrap();
     sleep_ms(100);
     // The resolved-ts must be blocked within lease.
     suite.must_get_rts(region.id, blocked_rts1);
     // Until we explicitly release the lease.
-    suite.must_release_sst_lease(region.id, &meta);
+    let import = suite.get_import_client(region.id);
+    must_release_sst_lease(import, &meta);
     suite.must_get_rts_ge(region.id, blocked_rts1);
 
     // Block resolved ts until lease expires.
-    suite.must_acquire_sst_lease(region.id, &meta, Duration::from_millis(100));
+    let import = suite.get_import_client(region.id);
+    must_acquire_sst_lease(import, &meta, Duration::from_millis(100));
     let blocked_rts2 = suite.region_resolved_ts(region.id).unwrap();
     sleep_ms(200);
     suite.must_get_rts_ge(region.id, blocked_rts2);
