+++ components/resolved_ts/tests/mod.rs
@@ -4,17 +4,8 @@ use std::{sync::*, time::Duration};
 
 use collections::HashMap;
 use engine_rocks::RocksEngine;
-use futures::{executor::block_on, stream, SinkExt};
-use grpcio::{ChannelBuilder, Environment, Result, WriteFlags};
-use kvproto::{
-    import_sstpb::{
-        AcquireLease, IngestRequest, LeaseRequest, ReleaseLease, SstMeta, UploadRequest,
-        UploadResponse,
-    },
-    import_sstpb_grpc::ImportSstClient,
-    kvrpcpb::*,
-    tikvpb::TikvClient,
-};
+use grpcio::{ChannelBuilder, Environment};
+use kvproto::{import_sstpb_grpc::ImportSstClient, kvrpcpb::*, tikvpb::TikvClient};
 use online_config::ConfigValue;
 use resolved_ts::Task;
 use test_raftstore::*;
@@ -252,84 +243,5 @@ impl TestSuite {
             sleep_ms(100)
         }
         panic!("fail to get greater ts after 50 trys");
-    }
-
-    pub fn must_acquire_sst_lease(&mut self, region_id: u64, meta: &SstMeta, ttl: Duration) {
-        let import = self.get_import_client(region_id);
-        let mut acquire = AcquireLease::default();
-        acquire.mut_lease().mut_region().set_id(region_id);
-        acquire.mut_lease().set_uuid(meta.get_uuid().into());
-        acquire.set_ttl(ttl.as_secs());
-        let mut req = LeaseRequest::default();
-        req.mut_acquire().push(acquire);
-
-        let resp = import.lease(&req).unwrap();
-
-        let acquired_lease = &resp.get_acquired()[0];
-        assert_eq!(region_id, acquired_lease.get_region().get_id(),);
-        assert_eq!(meta.get_uuid(), acquired_lease.get_uuid(),);
-    }
-
-    pub fn must_release_sst_lease(&mut self, region_id: u64, meta: &SstMeta) {
-        let import = self.get_import_client(region_id);
-        let mut release = ReleaseLease::default();
-        release.mut_lease().mut_region().set_id(region_id);
-        release.mut_lease().set_uuid(meta.get_uuid().into());
-        let mut req = LeaseRequest::default();
-        req.mut_release().push(release);
-
-        let resp = import.lease(&req).unwrap();
-
-        let released_lease = &resp.get_released()[0];
-        assert_eq!(region_id, released_lease.get_region().get_id(),);
-        assert_eq!(meta.get_uuid(), released_lease.get_uuid(),);
-    }
-
-    pub fn upload_sst(
-        &mut self,
-        region_id: u64,
-        meta: &SstMeta,
-        data: &[u8],
-    ) -> Result<UploadResponse> {
-        let import = self.get_import_client(region_id);
-        let mut r1 = UploadRequest::default();
-        r1.set_meta(meta.clone());
-        let mut r2 = UploadRequest::default();
-        r2.set_data(data.to_vec());
-        let reqs: Vec<_> = vec![r1, r2]
-            .into_iter()
-            .map(|r| Result::Ok((r, WriteFlags::default())))
-            .collect();
-        let (mut tx, rx) = import.upload().unwrap();
-        let mut stream = stream::iter(reqs);
-        block_on(async move {
-            let send_res = tx.send_all(&mut stream).await;
-            let close_res = tx.close().await;
-            match rx.await {
-                Ok(resp) => Ok(resp),
-                Err(e) => {
-                    send_res?;
-                    close_res?;
-                    Err(e)
-                }
-            }
-        })
-    }
-
-    pub fn must_ingest_sst(&mut self, region_id: u64, meta: SstMeta) {
-        let mut ingest_request = IngestRequest::default();
-        ingest_request.set_context(self.get_context(region_id));
-        ingest_request.set_sst(meta);
-
-        let ingest_sst_resp = self
-            .get_import_client(region_id)
-            .ingest(&ingest_request)
-            .unwrap();
-
-        assert!(
-            !ingest_sst_resp.has_error(),
-            "{:?}",
-            ingest_sst_resp.get_error()
-        );
     }
 }
