+++ tests/integrations/import/test_apply_log.rs
@@ -1,13 +1,16 @@
+// Copyright 2024 TiKV Project Authors. Licensed under Apache-2.0.
+
 use engine_traits::CF_DEFAULT;
 use external_storage::LocalStorage;
 use kvproto::import_sstpb::ApplyRequest;
 use tempfile::TempDir;
+use test_sst_importer::*;
 
-use crate::import::util;
+use super::util::*;
 
 #[test]
 fn test_basic_apply() {
-    let (_cluster, ctx, tikv, import) = util::new_cluster_and_tikv_import_client();
+    let (_cluster, ctx, tikv, import) = new_cluster_and_tikv_import_client();
     let tmp = TempDir::new().unwrap();
     let storage = LocalStorage::new(tmp.path()).unwrap();
     let default = [
@@ -17,20 +20,20 @@ fn test_basic_apply() {
         (b"k4", b"v4", 4),
     ];
     let default_rewritten = [(b"r1", b"v1", 1), (b"r2", b"v2", 2), (b"r3", b"v3", 3)];
-    let mut sst_meta = util::make_plain_file(&storage, "file1.log", default.into_iter());
-    util::register_range_for(&mut sst_meta, b"k1", b"k3a");
+    let mut sst_meta = make_plain_file(&storage, "file1.log", default.into_iter());
+    register_range_for(&mut sst_meta, b"k1", b"k3a");
     let mut req = ApplyRequest::new();
     req.set_context(ctx.clone());
-    req.set_rewrite_rules(vec![util::rewrite_for(&mut sst_meta, b"k", b"r")].into());
+    req.set_rewrite_rules(vec![rewrite_for(&mut sst_meta, b"k", b"r")].into());
     req.set_metas(vec![sst_meta].into());
-    req.set_storage_backend(util::local_storage(&tmp));
+    req.set_storage_backend(local_storage(&tmp));
     import.apply(&req).unwrap();
-    util::check_applied_kvs_cf(&tikv, &ctx, CF_DEFAULT, default_rewritten.into_iter());
+    check_applied_kvs_cf(&tikv, &ctx, CF_DEFAULT, default_rewritten.into_iter());
 }
 
 #[test]
 fn test_apply_twice() {
-    let (_cluster, ctx, tikv, import) = util::new_cluster_and_tikv_import_client();
+    let (_cluster, ctx, tikv, import) = new_cluster_and_tikv_import_client();
     let tmp = TempDir::new().unwrap();
     let storage = LocalStorage::new(tmp.path()).unwrap();
     let default = [(
@@ -49,21 +52,21 @@ fn test_apply_twice() {
         1,
     )];
 
-    let mut sst_meta = util::make_plain_file(&storage, "file2.log", default.into_iter());
-    util::register_range_for(&mut sst_meta, b"k1", b"k1a");
+    let mut sst_meta = make_plain_file(&storage, "file2.log", default.into_iter());
+    register_range_for(&mut sst_meta, b"k1", b"k1a");
     let mut req = ApplyRequest::new();
     req.set_context(ctx.clone());
-    req.set_rewrite_rules(vec![util::rewrite_for(&mut sst_meta, b"k", b"r")].into());
+    req.set_rewrite_rules(vec![rewrite_for(&mut sst_meta, b"k", b"r")].into());
     req.set_metas(vec![sst_meta.clone()].into());
-    req.set_storage_backend(util::local_storage(&tmp));
+    req.set_storage_backend(local_storage(&tmp));
     import.apply(&req).unwrap();
-    util::check_applied_kvs_cf(&tikv, &ctx, CF_DEFAULT, default_fst.into_iter());
+    check_applied_kvs_cf(&tikv, &ctx, CF_DEFAULT, default_fst.into_iter());
 
-    util::register_range_for(&mut sst_meta, b"k1", b"k1a");
-    req.set_rewrite_rules(vec![util::rewrite_for(&mut sst_meta, b"k", b"z")].into());
+    register_range_for(&mut sst_meta, b"k1", b"k1a");
+    req.set_rewrite_rules(vec![rewrite_for(&mut sst_meta, b"k", b"z")].into());
     req.set_metas(vec![sst_meta].into());
     import.apply(&req).unwrap();
-    util::check_applied_kvs_cf(
+    check_applied_kvs_cf(
         &tikv,
         &ctx,
         CF_DEFAULT,
